【AC】[AtCoder Grand Contest 023]A - Zero-Sum Ranges

+++ agc023/agc023_a.cpp
@@ -79,7 +79,7 @@ int main() {
         cs.set(i, a);
     }
     cs.calculate();
-    map<ll, int> m;
+    map<ll, ll> m;
     m[0] = 1;
     rep(i, n) m[cs.getSum(i)]++;
